SPSH-738: Fixed tests for the PersonenkontextScope
